requirements.txt
streamlit
pandas
matplotlib
seaborn
scipy
scikit-learn
Pillow
